refactor: extract Horizon Huntarr into recommenders/horizon.py (2.10.43)
PR2 step 2 of the external.py architecture decomposition - pure relocation of find_horizon_movies + its cache helpers. horizon.py imports get_collection_details/load_huntarr_cache from huntarr.py (deliberate sibling reuse, unchanged from before this move). external.py re-exports everything external callers/tests still need. Verified byte-identical against the golden-output harness.
